backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pymupdf==1.25.2
anthropic==0.38.0
python-pptx==1.0.2
pydantic==2.9.2
